% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_gen.R
\name{ar_gen}
\alias{ar_gen}
\title{Generate an autoregressive model.}
\usage{
ar_gen(p, phi_vec)
}
\arguments{
\item{p}{the dimension of L}

\item{phi_vec}{a K-dimensional vector for off-diagonal values}
}
\value{
a p-by-p strictly banded lower triangular matrix
}
\description{
Generate lower triangular matrix with strict bandwidth. See, e.g., Model 1 in the paper.
}
\examples{
true_ar <- ar_gen(p = 50, phi = c(0.5, -0.4, 0.1))
}

