% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_gen.R
\name{block_diag_gen}
\alias{block_diag_gen}
\title{Generate a model with block-diagonal structure}
\usage{
block_diag_gen(p)
}
\arguments{
\item{p}{the dimension of L}
}
\value{
a p-by-p lower triangular matrix with block-diagonal structure from p/4-th row to 3p/4-th row
}
\description{
Generate a model with block-diagonal structure
}
\examples{
set.seed(123)
true_L_block_diag <- block_diag_gen(p = 50)
}

