% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matimage}
\alias{matimage}
\title{Plot the sparsity pattern of a square matrix}
\usage{
matimage(Mat, main = NULL)
}
\arguments{
\item{Mat}{A matrix to plot.}

\item{main}{A plot title.}
}
\description{
Black, white and gray stand for positive, zero and negative respectively
}
\examples{
set.seed(123)
p <- 50
n <- 50
phi <- 0.4
true <- varband_gen(p = p, block = 5)
matimage(true)
}

